src/main/java/com/mnubo/java/sdk/client/services/EventsSDKServices.java
package com.mnubo.java.sdk.client.services;

import static com.mnubo.java.sdk.client.Constants.OBJECT_PATH;
import static com.mnubo.java.sdk.client.utils.ValidationUtils.notBlank;
import static com.mnubo.java.sdk.client.utils.ValidationUtils.notNullNorEmpty;
import static com.mnubo.java.sdk.client.utils.ValidationUtils.validNotNull;
import static java.util.Arrays.*;

import java.util.*;

import com.mnubo.java.sdk.client.models.Event;
import com.mnubo.java.sdk.client.models.result.Result;
import com.mnubo.java.sdk.client.spi.EventsSDK;

class EventsSDKServices implements EventsSDK {

    public static final String EVENT_PATH = "/events";
    public static final String EVENT_PATH_SEGMENT = "events";
    public static final String EVENT_PATH_EXITS = EVENT_PATH + "/exists";
    private static final String REPORT_RESULTS_QUERY_PARAM = "report_results";
    private final SDKService sdkCommonServices;

    EventsSDKServices(SDKService sdkCommonServices) {
        this.sdkCommonServices = sdkCommonServices;
    }

    @Override
    public List<Result> send(String deviceId, List<Event> events) {

        notBlank(deviceId, "x_device_Id cannot be blank.");
        notNullNorEmpty(events, "Event list cannot be null or empty.");

        final String url = sdkCommonServices.getIngestionBaseUri()
                                            .path(OBJECT_PATH)
                                            .pathSegment(deviceId, EVENT_PATH_SEGMENT)
                                            .queryParam(REPORT_RESULTS_QUERY_PARAM, true)
                                            .build().toString();

        return postRequest(url, events);
    }



    @Override
    public List<Result> send(List<Event> events) {

        notNullNorEmpty(events, "Event list cannot be null or empty.");

        final String url = sdkCommonServices.getIngestionBaseUri()
                                            .path(EVENT_PATH)
                                            .queryParam(REPORT_RESULTS_QUERY_PARAM, true)
                                            .build().toString();

        return postRequest(url, events);
    }

    @Override
    public List<Result> send(Event... events) {

        return send(asList(events));
    }

    @Override
    public List<Map<String, Boolean>> eventsExist(List<UUID> eventIds) {
        validNotNull(eventIds, "List of the eventIds cannot be null.");

        final String url = sdkCommonServices.getIngestionBaseUri()
                .path(EVENT_PATH_EXITS)
                .build().toString();

        List<Map<String,Boolean>> result = new ArrayList<>();
        return sdkCommonServices.postRequest(url, result.getClass(), eventIds);
    }

    @Override
    public Boolean isEventExists(UUID eventId) {
        validNotNull(eventId, "eventId cannot be blank.");

        final String url = sdkCommonServices.getIngestionBaseUri()
                .path(EVENT_PATH_EXITS)
                .pathSegment(eventId.toString())
                .build().toString();

        Map<String, Boolean> results = new HashMap<>();
        results = sdkCommonServices.getRequest(url, results.getClass());
        return results == null || results.size() != 1 || results.get(eventId.toString()) == null ?
                false : results.get(eventId.toString());
    }

    private List<Result> postRequest(String url, Object object) {
        Result[] results = sdkCommonServices.postRequest(url, Result[].class, object);
        return results == null ? new ArrayList<Result>() : Arrays.asList(results);
    }
}
